requirements.txt
# Core data science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0.0

# Bioinformatics
biopython>=1.79

# Deep learning
torch>=1.10.0
torchvision>=0.11.0
torchaudio>=0.10.0

# Optional: for graph neural networks
# torch-geometric>=2.0.0

# Kaggle
kaggle>=1.5.0

# Utilities
tqdm>=4.62.0
pyyaml>=6.0